feat: [MR-523] Prevent enqueueing of multiple responses for the same callback
[MR-523] Currently, when inducting a response we only check that it matches an
existing callback, dropping any mismatched responses while raising a critical
error. However, if due to malicious behavior or bugs, a second response for the
same callback is produced before the first one has been executed, this second
response would also get enqueued and consume an extra slot reservation; which
would later result in dropping a legitimate response due to it not having a
response slot reservation.

We introduce an explicit, transient set of callbacks IDs whose responses are
already enqueued in an input queue; and check against this set for potential
duplicate responses. We return an error if this is a guaranteed response (which
results in a critical error); but silently ignore duplcate best-effort responses
(as this may be a response delivered after we have already produced a local
timeout reject response).

This functionality is required for safely generating local timeout reject
responses for expired best-effort callbacks.