feat: [MR-603] Keep track of shed inbound responses
Inbound best-effort responses may be shed due to memory pressure. But we still
want to produce a `SYS_UNKNOWN` reject response for the respective `CallbackId`.

So we shed the response from the pool, but remember it (and its `CallbackId`) in
a "dropped inbound responses" map. Input queue references are non-stale if they
point either into the `MessagePool` or into the "dropped inbound responses" map.
When we pop such a reference from an input queue, `CanisterQueues` returns a
`CanisterInput::UnknownResponse(CallbackId)` (as opposed to the usual
`CanisterInput::Response(Arc<Response>)`; this is then inflated by `SystemState`
(with the help of its `CallContextManager`) into a perfectly matching reject
response.